fix: preflight pi runtime before database startup
