+++ lib/setup.mjs
@@ -14,7 +14,7 @@ export const SETUP_CLIENTS = Object.freeze([
   { id: "antigravity", name: "Antigravity CLI", executable: "agy" },
 ]);
 const SOURCE = fileURLToPath(new URL("../", import.meta.url));
-const RUNTIME_FILES = ["package.json", "extension.mjs", "lore-pi.ts", "lore-server.mjs", "lore-cli.mjs", "pi-session-reader.mjs", "lib", "schemas", "browser", "scripts", "lore.example.json"];
+const RUNTIME_FILES = ["package.json", "extension.mjs", "lore-pi.ts", "lore-server.mjs", "lore-server-runtime.mjs", "lore-cli.mjs", "pi-session-reader.mjs", "lib", "schemas", "browser", "scripts", "lore.example.json"];
 
 function targetStat(target) {
   try { return lstatSync(target); } catch (error) { if (error.code === "ENOENT") return null; throw error; }
