+++ tests/fixtures/block-sqlite-loader.mjs
@@ -1,0 +1,6 @@
+export async function resolve(specifier, context, nextResolve) {
+  if (specifier === "node:sqlite") {
+    throw new Error("blocked node:sqlite for runtime preflight test");
+  }
+  return nextResolve(specifier, context);
+}
